renamed cooperate folder to backend
