+++ backend/src/main/java/com/example/cooperate/DataTransferObject.java
@@ -1,0 +1,6 @@
+package com.example.cooperate;
+
+public interface DataTransferObject {
+
+    int getId();
+}
