--- cooperate/src/main/java/com/example/cooperate/DataTransferObject.java
@@ -1,6 +1,0 @@
-package com.example.cooperate;
-
-public interface DataTransferObject {
-
-    int getId();
-}
